pyproject.toml
[project]
name = "spypixel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "fastapi>=0.116.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "user-agents>=2.2.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
